pyproject.toml
[project]
name = "jptv-subtitles"
version = "0.1.0"
description = "jptv subtitle downloader"
authors = [
    {name = "markgrandi@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lxml (>=5.3.1,<6.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "aiohttp (>=3.11.13,<4.0.0)",
    "attrs (>=25.3.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
